Combine GIT_LFS_PROGRESS logger with ProgressMeter
